More removal, adjust tests
