+++ webapp/routes.py
@@ -1,7 +1,7 @@
 from datetime import datetime as dt
 import logging
 from flask import render_template, redirect
-from webapp import app, pgconfig2
+from webapp import app, pgconfig
 
 
 LOGGER = logging.getLogger(__name__)
@@ -38,7 +38,7 @@ def view_app_param_not_set():
 @app.route('/param/<pg_param>')
 def view_app_params(pg_param):
     select_html = get_param_select_html(pg_param)
-    param_details = pgconfig2.get_pg_param_over_versions(pg_param)
+    param_details = pgconfig.get_pg_param_over_versions(pg_param)
     return render_template('param2.html',
                            year=get_year(),
                            param_details=param_details,
@@ -48,8 +48,8 @@ def view_app_params(pg_param):
 
 @app.route('/param/change/<vers1>/<vers2>')
 def view_app_param_changes_v2(vers1, vers2):
-    vers1_redirect = pgconfig2.check_redirect(version=vers1)
-    vers2_redirect = pgconfig2.check_redirect(version=vers2)
+    vers1_redirect = pgconfig.check_redirect(version=vers1)
+    vers2_redirect = pgconfig.check_redirect(version=vers2)
 
     if vers1 == vers2:
         return redirect('/param/change')
@@ -61,12 +61,12 @@ def view_app_param_changes_v2(vers1, vers2):
     vers2_html = _version_select_html(name='version_2', filter_default=vers2)
 
     try:
-        config_changes = pgconfig2.config_changes(vers1, vers2)
+        config_changes = pgconfig.config_changes(vers1, vers2)
     except ValueError:
         return redirect('/param/change')
 
-    config_changes_html = pgconfig2.config_changes_html(config_changes)
-    changes_stats = pgconfig2.config_changes_stats(config_changes)
+    config_changes_html = pgconfig.config_changes_html(config_changes)
+    changes_stats = pgconfig.config_changes_stats(config_changes)
     return render_template('param_change2.html',
                            year=get_year(),
                            config_changes=config_changes_html,
@@ -119,7 +119,7 @@ def get_param_select_html(filter_default: str='max_parallel_workers_per_gather')
     html : str
     """
     html = '<select id="selection_param" name="selection_param"> '
-    options = pgconfig2.get_all_postgres_parameters()
+    options = pgconfig.get_all_postgres_parameters()
     options.sort()
     for option in options:
         if option == filter_default:
@@ -136,7 +136,7 @@ def get_param_select_html(filter_default: str='max_parallel_workers_per_gather')
 
 def _version_select_html(name, filter_default):
     html = '<select id="{name}" name="{name}"> '.format(name=name)
-    options = pgconfig2.VERSIONS
+    options = pgconfig.VERSIONS
     for option in options:
         if option == filter_default:
             tmp_html = '<option value="{}" selected="selected">{}</option>'
